remote: move foreign layer SSRF helpers to fetcher.go per review
Move validateForeignURL and fetchForeignBlobURL (plus their tests) next
to the fetcher receiver, trim verbose function comments, and rework
Compressed() to call fetchBlobURL on the single registry URL with
foreign URLs collected and iterated separately.
