+++ pkg/v1/remote/image_test.go
@@ -749,196 +749,3 @@ func TestData(t *testing.T) {
 		t.Fatal(err)
 	}
 }
-
-func TestValidateForeignURL(t *testing.T) {
-	tests := []struct {
-		url      string
-		insecure bool
-		wantErr  bool
-	}{
-		// public HTTPS — always allowed
-		{"https://cdn.example.com/layer.tar.gz", false, false},
-		{"https://8.8.8.8/layer.tar.gz", false, false},
-		// public HTTP — only for insecure registries
-		{"http://cdn.example.com/layer.tar.gz", true, false},
-		{"http://cdn.example.com/layer.tar.gz", false, true},
-		// loopback — rejected even for insecure
-		{"http://127.0.0.1/layer.tar.gz", true, true},
-		{"https://127.0.0.1/layer.tar.gz", false, true},
-		{"https://[::1]/layer.tar.gz", false, true},
-		// link-local (cloud metadata)
-		{"https://169.254.169.254/latest/meta-data/", false, true},
-		{"http://169.254.169.254/latest/meta-data/", true, true},
-		// private RFC 1918
-		{"https://10.0.0.1/layer.tar.gz", false, true},
-		{"https://192.168.1.1/layer.tar.gz", false, true},
-		// unspecified
-		{"https://0.0.0.0/layer.tar.gz", false, true},
-		// disallowed schemes
-		{"ftp://cdn.example.com/layer.tar.gz", false, true},
-		{"file:///etc/passwd", false, true},
-	}
-	for _, tt := range tests {
-		err := validateForeignURL(tt.url, tt.insecure)
-		if tt.wantErr && err == nil {
-			t.Errorf("validateForeignURL(%q, insecure=%v) should have been rejected", tt.url, tt.insecure)
-		}
-		if !tt.wantErr && err != nil {
-			t.Errorf("validateForeignURL(%q, insecure=%v) unexpected error: %v", tt.url, tt.insecure, err)
-		}
-	}
-}
-
-// TestPullingForeignLayerSSRF verifies that a manifest whose foreign-layer URL
-// points to a private or loopback address is rejected before any network
-// request is made, preventing SSRF.
-func TestPullingForeignLayerSSRF(t *testing.T) {
-	img := randomImage(t)
-	expectedRepo := "foo/bar"
-
-	foreignLayer, err := random.Layer(1024, types.DockerForeignLayer)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Embed a foreign layer URL pointing to the AWS metadata endpoint.
-	img, err = mutate.Append(img, mutate.Addendum{
-		Layer: foreignLayer,
-		URLs:  []string{"http://169.254.169.254/latest/meta-data/iam/security-credentials/"},
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	manifestPath := fmt.Sprintf("/v2/%s/manifests/latest", expectedRepo)
-	configPath := fmt.Sprintf("/v2/%s/blobs/%s", expectedRepo, mustConfigName(t, img))
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		switch r.URL.Path {
-		case "/v2/":
-			w.WriteHeader(http.StatusOK)
-		case configPath:
-			w.Write(mustRawConfigFile(t, img))
-		case manifestPath:
-			w.Write(mustRawManifest(t, img))
-		default:
-			t.Fatalf("unexpected path: %v", r.URL.Path)
-		}
-	}))
-	defer server.Close()
-	u, err := url.Parse(server.URL)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	tag := mustNewTag(t, fmt.Sprintf("%s/%s:latest", u.Host, expectedRepo))
-	rmt, err := Image(tag, WithTransport(http.DefaultTransport))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Pulling the foreign layer must be rejected.
-	layers, err := rmt.Layers()
-	if err != nil {
-		t.Fatal(err)
-	}
-	_, err = layers[1].Compressed()
-	if err == nil {
-		t.Error("Compressed() should have been rejected for a foreign layer URL pointing to a private address")
-	}
-}
-
-// TestPullingForeignLayerSSRFViaRedirect verifies that the redirect-based SSRF
-// bypass is closed.  Without the CheckRedirect guard in fetchForeignBlobURL, an
-// attacker-controlled CDN host that the initial validateForeignURL check passes
-// could issue a 302 to a private/loopback address (e.g. 169.254.169.254) and
-// the Go HTTP client would follow it, leaking cloud-instance credentials.
-//
-// Attack model:
-//  1. Attacker publishes an OCI image whose manifest contains a foreign layer
-//     with URLs: ["https://cdn.attacker.com/layer.tar.gz"].
-//  2. The initial validateForeignURL check passes (cdn.attacker.com is public).
-//  3. cdn.attacker.com returns HTTP 302 → http://169.254.169.254/credentials.
-//  4. Without the fix the Go HTTP client follows the redirect and returns
-//     cloud IAM credentials to the caller.
-func TestPullingForeignLayerSSRFViaRedirect(t *testing.T) {
-	img := randomImage(t)
-	expectedRepo := "foo/bar"
-
-	foreignLayer, err := random.Layer(1024, types.DockerForeignLayer)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Victim: simulates a private metadata/credential server reachable on loopback.
-	victim := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		w.Write([]byte(`{"AccessKeyId":"ASIA_LEAKED","SecretAccessKey":"LEAKED_SECRET"}`))
-	}))
-	defer victim.Close()
-
-	// Attacker CDN: looks like a public server but immediately redirects to victim.
-	attacker := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		http.Redirect(w, r, victim.URL+"/credentials", http.StatusFound)
-	}))
-	defer attacker.Close()
-
-	// Embed the attacker URL as a foreign layer source.  The URL passes
-	// validateForeignURL (it is an http loopback URL for an insecure registry
-	// test — treated as insecure below), but the redirect leads to another
-	// loopback address that must be rejected.
-	img, err = mutate.Append(img, mutate.Addendum{
-		Layer: foreignLayer,
-		URLs:  []string{attacker.URL + "/layer.tar.gz"},
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	manifestPath := fmt.Sprintf("/v2/%s/manifests/latest", expectedRepo)
-	configPath := fmt.Sprintf("/v2/%s/blobs/%s", expectedRepo, mustConfigName(t, img))
-
-	registryServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		switch r.URL.Path {
-		case "/v2/":
-			w.WriteHeader(http.StatusOK)
-		case configPath:
-			w.Write(mustRawConfigFile(t, img))
-		case manifestPath:
-			w.Write(mustRawManifest(t, img))
-		default:
-			// Return 404 so the client falls through to the foreign layer URL.
-			w.WriteHeader(http.StatusNotFound)
-		}
-	}))
-	defer registryServer.Close()
-
-	u, err := url.Parse(registryServer.URL)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Use WithInsecure so the http:// attacker URL passes scheme validation.
-	ref, err := name.ParseReference(fmt.Sprintf("%s/%s:latest", u.Host, expectedRepo), name.Insecure)
-	if err != nil {
-		t.Fatal(err)
-	}
-	rmt, err := Image(ref, WithTransport(http.DefaultTransport))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	layers, err := rmt.Layers()
-	if err != nil {
-		t.Fatal(err)
-	}
-	// The foreign layer fetch must fail because the redirect points to a loopback address.
-	_, err = layers[1].Compressed()
-	if err == nil {
-		t.Error("Compressed() followed a redirect to a private address — SSRF bypass not blocked")
-	}
-	if err != nil && !strings.Contains(err.Error(), "private or link-local") {
-		// The error might also appear wrapped; as long as no credentials are returned it is safe.
-		t.Logf("Compressed() returned error (expected): %v", err)
-	}
-}
